server/requirements.txt
fastapi
uvicorn[standard]
websockets
sqlalchemy
python-multipart
bcrypt==3.2.0
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
